Fix outdated GraphQL query, and attempt to introduce pagination
NAIS API just released v1, and as a result of that the queries in
"Teamkatalogen" must be updated. v1 of NAIS API implements the Cursor
Connections Specification instead of the limit / offset based pagination
we previously had.

Basically we fetch a page of teams, and the pageInfo object holds
information about how to continue to fetch more.

The query for single teams has not changed in v1, so no changes needed
to be made to that.

Also, the GraphQL endpoint has been changed from "/query" to "/graphql".

More info:

- https://github.com/nais/api/blob/main/docs/graphql_practices.md#pagination
- https://relay.dev/graphql/connections.htm